backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
motor==3.7.0
pymongo==4.12.1
pydantic==2.11.3
pydantic-settings==2.9.1
python-dotenv==1.1.0
librosa==0.11.0
numpy>=2.0.0
soundfile==0.13.1
pydub==0.25.1
aiofiles==24.1.0
httpx==0.28.1
scikit-learn>=1.5.0
